Resolve mounts' host_path relatively to vmtest.toml
When a non-default `rootfs` is used, and vmtest is run through
`go run|test -exec`, the binary executed within the VM can't be found,
because that rootfs doesn't contain the test binary compiled on the fly.
So a custom mount has to be specified.

But, before that commit, if a relative `host_path`` was used, it'd be
passed as is to qemu, and thus would be resolved as relative to the cwd
where qemu was running.

In the case of Go, if a directory was passed to `go test`, eg.
`./integration/...`, then Go would switch to the directory under test,
and then run its exec argument. But, the right way to use `go test -run`
with `vmtest` is to set `GOTMPDIR=.` to write the test binary within
the directory where `go test -exec` is run.

To work around this mismatch, resolve `host_path` relative to the
`vmtest.toml` file.

FWIW this doesn't fully solve the original issue, unless the VM path
put in the config file is hardcoded to the cwd of `go test -exec`. To
overcome this, one can prepend the inline command with a call to `ln`:

```sh
GOTMPDIR=. go test -exec "vmtest -c $(pwd)/vmtest.toml -- ln -s /skbdump/ $(pwd); " ./integration/...
```

Signed-off-by: Albin Kerouanton <albinker@gmail.com>